Appending null terminators to the strings while passing to NI's library.
Resolving an issue wherein passed-in strings are expected by NI driver to be null-terminated, and due to garbage values the calls fail occasionally. Therefore, the bytes from getBytes() are taken as an array after appending the strings with a space character, and making it zero after getting the bytes of string so as to pass null terminated strings to NI's library.